requirements.txt
numpy==1.14.0
Pillow==4.1.1
tracery==0.1.1